Split out `infalliable_detructuring_match`
